Extract image pickers + dedupe and bg processing of images
